+++ changelog.md
@@ -6,6 +6,7 @@
 
 * JSDoc of all public classes and methods.
 * New util functions `generateRandomFctAddress` and `generateRandomEcAddress`.
+* Method `toObject()` added to `Entry` and `Chain` classes. Returns a JavaScript object than can be serialized.
 
 ### Fixed